Wanted to make a basic GUI, ended up spending 4 days debugging and wracking my brain over why nothing is working. Some bug fixes, mostly overhaul of graphics and im still not happy with it. Basic window system whatever. GRAPHICS ESCAPADE :D

+++ kernel/src/drivers/mouse.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 #include <drivers/ps2.h>
 
-typedef void (* mouse_handler_t)(int32_t, int32_t, bool[3]);
+typedef void (* mouse_handler_t)(int16_t, int16_t, bool[3]);
 
 void mouse_initialize(ps2_ports_t port);
 void mouse_set_handler(mouse_handler_t mouse_handler);
